chore(ad-planning): move match-custom-goal helper out of subscribers/ + drop empty inventory-created
Medusa's subscriber loader recursively scans src/subscribers/ and warns
"is not a function. skipped" for any file without a default handler export.
Two files tripped it at boot:
- match-custom-goal.ts (a pure helper added in #615) → moved to the ad-planning
  module's utils/ (with its unit test), where pure domain helpers belong
  (mirrors partner_billing/lib/). Import in analytics-event-created updated.
- inventory-created.ts (0 bytes, dead) → deleted.

No behaviour change; matcher unit test (5) still green at the new path.
